added object to move on trajectory
