+++ statusmodifiers.py
@@ -1,8 +1,8 @@
 from moveableobject import MoveableObject
+from trajectorymovingobject import TrajectoryMovingObject
 from observe import Event
 from constants import SCREENW, SCREENH, STATMOD_DURATION, PLAYERSPEED, PLAYERMAXSPEED, STATMOD_SPEED
 from timer import Timer
-import math
 import random
 
 
@@ -12,36 +12,17 @@ import random
 # the first power up received
 
 # power ups and downs, to be inherited from
-class StatusModifier(MoveableObject):
+class StatusModifier(TrajectoryMovingObject):
     def __init__(self, img):
-        MoveableObject.__init__(self, random.randrange(0, SCREENW), -100, STATMOD_SPEED, img)
+        TrajectoryMovingObject.__init__(self, random.randrange(0, SCREENW), -1 * img.get_height(), STATMOD_SPEED, img)
         # angle randomly ranges from 100 degrees to 260, 0 degrees is vertical axis
-        self.degreeangle = random.randrange(100, 260)
         self.timer = Timer(self)
 
     def payload(self, target):
         self.notify(Event("remove"))
 
-    def move(self):
-        # screen edge checking
-
-        if (self.x > SCREENW - self.width and self.degreeangle > 180) or random.randrange(0, 2000) == 1467:
-            self.degreeangle = random.randrange(100, 160)  # 270 - (self.degreeangle - 90)
-        # print "turn left"
-        elif (self.x < 0 and self.degreeangle < 180) or random.randrange(0, 2000) == 200:
-            self.degreeangle = random.randrange(200, 260)  # 90 + (270 - self.degreeangle)
-        # print "turn right"
-
-        self.x -= (((math.degrees(math.sin(math.radians(self.degreeangle))) * self.speed) / 40) * self.frame_tick)
-        self.y -= (((math.degrees(math.cos(math.radians(self.degreeangle))) * self.speed) / 40) * self.frame_tick)
-        self.updatepos()
-
-    # print "x = " + str(self.x) + ", y = " + str(self.y)
-    # create movement trigonometrically like in Panzer Deathmatch
-
     def update(self):
         super().update()
-        self.move()
         if self.y > SCREENH:
             self.notify(Event("remove"))
